// src/interfaces/interfaces.ts
export interface InputData {
  age: number;
  heigth: number;
  hairColor: "brown" | "black" | "blonde" | "redhead" | "grey";
  hairHeigth: "large" | "middle" | "short";
  eyeColor: "black" | "brown" | "blue" | "green" | "grey";
  boobSize: "small" | "middle" | "big" | "veryBig";
  physicalContext: "normal" | "thin" | "sporty" | "chubby" | "fat";
}